+++ src/HealthCalculator.php
@@ -4,5 +4,5 @@ namespace Tomfordweb\HealthCalculators;
 
 interface HealthCalculator
 {
-    public function calculate(): int;
+    public function calculate(): float;
 }
